+++ crates/structs/src/steam.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Deserialize, Serialize, Debug)]
 pub struct App{
     #[serde(rename = "appid")]
-    pub app_id: usize,
+    pub app_id: u32,
     pub name: String,
     pub last_modified: i64,
     pub price_change_number: i64,
@@ -11,7 +11,7 @@ pub struct App{
 
 pub struct PriceOverview{
     pub currency: String,
-    pub discount_percent: usize,
+    pub discount_percent: u32,
     pub initial: f64,
     pub final_price: f64,
 }
